[Refactor] modified List component with composition pattern
