+++ src/components/List/ListItem.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Image from 'next/image';
 import Star from '@/components/Star';
 import { truncStr } from '@/utils/truncStr';
